Add files via upload
The file can be run by jupyter anaconda. The webdriver can be downloaded online. You have to set the path of the webdriver in the 'string name a' in line 3 of this file.
The program will automatically open spie.org and proceed with the form fillup with provided data. You might need to change the data for actual results. The data provided here are not real and is based for example purpose. It is advisory to check the data before opting for membership payment.